# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/doctors": {
      "post": {
        "operationId": "DoctorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "503": {
            "description": "Error creating doctor"
          }
        },
        "tags": [
          "Doctors"
        ]
      },
      "get": {
        "operationId": "DoctorsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "specialization_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "professional_since_from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "professional_since_to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "503": {
            "description": "Error finding all doctors"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
    "/doctors/me": {
      "get": {
        "operationId": "DoctorsController_findDoctorsProfileOfUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error finding doctor"
          }
        },
        "tags": [
          "Doctors"
        ]
      },
      "patch": {
        "operationId": "DoctorsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoctorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "400": {
            "description": "Specialization not found"
          },
          "403": {
            "description": "Doctor profile id and user id do not match"
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error updating doctor"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
    "/doctors/{id}": {
      "get": {
        "operationId": "DoctorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error finding doctor"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
    "/doctors/{doctorId}/templates": {
      "get": {
        "operationId": "DoctorsController_getDoctorTemplates",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "503": {
            "description": "Error finding template"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
    "/doctors/templates": {
      "post": {
        "operationId": "DoctorsController_createTemplateSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error creating template"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
    "/doctors/ratings": {
      "post": {
        "operationId": "DoctorsController_createDoctorRating",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorsRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorsRatingDto"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found"
          },
          "503": {
            "description": "Error creating doctors rating"
          }
        },
        "tags": [
          "Doctors"
        ]
      }
    },
...
    "/appointments/doctors/{doctorId}": {
      "get": {
        "operationId": "AppointmentsController_findByDoctorId",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Error finding appointments"
          }
        },
        "tags": [
          "Appointments"
        ]
      }
    },
...
    "/timeslots/doctors/{doctorId}": {
      "get": {
        "operationId": "TimeslotsController_getDoctorTimeslots",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeslotDto"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Error finding timeslots"
          }
        },
        "tags": [
          "Timeslots"
        ]
      }
    },
...
  },
  "info": {
    "title": "Appointments service API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateDoctorDto": {
        "type": "object",
        "properties": {
          "phone_number": {
            "type": "string"
          },
          "licence_number": {
            "type": "string"
          },
          "specializationId": {
            "type": "number",
            "minimum": 1
          },
          "hospital_address": {
            "type": "object"
          },
          "hospital_name": {
            "type": "object"
          },
          "professional_since": {
            "type": "object"
          }
        },
        "required": [
          "phone_number",
          "licence_number",
          "specializationId",
          "hospital_address",
          "hospital_name",
          "professional_since"
        ]
      },
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "phone_number": {
            "type": "string"
          },
          "licence_number": {
            "type": "string"
          },
          "specialization_id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "phone_number",
          "licence_number",
          "specialization_id",
          "user_id"
        ]
      },
      "UpdateDoctorDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTemplateScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDoctorsRatingDto": {
        "type": "object",
        "properties": {
          "doctor_id": {
            "type": "number",
            "minimum": 1
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "review": {
            "type": "string"
          }
        },
        "required": [
          "doctor_id",
          "rating",
          "review"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "number"
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "date_of_birth",
          "gender",
          "address",
          "created_at",
          "updated_at"
        ]
      },
      "DoctorsRatingDto": {
        "type": "object",
        "properties": {
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "id": {
            "type": "number"
          },
          "rating": {
            "type": "number"
          },
          "review": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "doctor",
          "patient",
          "id",
          "rating",
          "review"
        ]
      },
      "TimeslotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "start_time": {
            "format": "date-time",
            "type": "string"
          },
          "end_time": {
            "format": "date-time",
            "type": "string"
          },
          "is_available": {
            "type": "boolean"
          },
          "doctor_id": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "start_time",
          "end_time",
          "is_available",
          "doctor_id"
        ]
      },
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "patient_insurance_number": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "timeslot_id": {
            "type": "number"
          },
          "doctor_id": {
            "type": "number"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "timeslot": {
            "$ref": "#/components/schemas/TimeslotDto"
          }
        },
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "patient_insurance_number",
          "reason",
          "timeslot_id",
          "doctor_id",
          "patient",
          "doctor",
          "timeslot"
        ]
      },
...
    }
  }
}